54: Figured out best way to handle mocks
