+++ rag-toolchain/src/clients/traits.rs
@@ -28,3 +28,31 @@ pub trait AsyncChatClient {
         prompt_messages: Vec<PromptMessage>,
     ) -> Result<PromptMessage, Self::ErrorType>;
 }
+
+#[cfg(test)]
+use mockall::*;
+#[cfg(test)]
+mock! {
+    pub AsyncEmbeddingClient {}
+    #[async_trait]
+    impl AsyncEmbeddingClient for AsyncEmbeddingClient {
+        type ErrorType = std::io::Error;
+        async fn generate_embedding(&self, text: Chunk) -> Result<(Chunk, Embedding), <Self as AsyncEmbeddingClient>::ErrorType>;
+        async fn generate_embeddings(
+            &self,
+            text: Chunks,
+        ) -> Result<Vec<(Chunk, Embedding)>, <Self as AsyncEmbeddingClient>::ErrorType>;
+    }
+}
+#[cfg(test)]
+mock! {
+    pub AsyncChatClient {}
+    #[async_trait]
+    impl AsyncChatClient for AsyncChatClient {
+        type ErrorType = std::io::Error;
+        async fn invoke(
+            &self,
+            prompt_messages: Vec<PromptMessage>,
+        ) -> Result<PromptMessage, <Self as AsyncChatClient>::ErrorType>;
+    }
+}
