+++ rag-toolchain/src/retrievers/mod.rs
@@ -5,3 +5,7 @@ pub use postgres_vector_retriever::{
     DistanceFunction, PostgresRetrieverError, PostgresVectorRetriever,
 };
 pub use traits::AsyncRetriever;
+
+// export the trait mocks for use in testing
+#[cfg(test)]
+pub use traits::MockAsyncRetriever;
